local_dev/backend/requirements.txt
# Backend Requirements for VANA Local Development

# Core Dependencies
python-dotenv>=1.0.0
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
google-auth>=2.23.0
google-api-python-client>=2.100.0

# Agent Dependencies
google-adk>=0.5.0
langchain>=0.0.267
langchain-core>=0.0.10
langchain-community>=0.0.10

# Vector Search
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Document Processing
PyPDF2>=3.0.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0

# Utilities
pyyaml>=6.0.0
tqdm>=4.66.0
colorlog>=6.7.0
